Created Unity Project
Created a Unity project and added some objects we'll need to have so we can create our features in branches.

Main Menu - Features a MainMenuCanvas new game, load game, options, and quit button ,

Test Scene - The main scene to do all our work in. This features a UICanvas object for all our UI to go in, and a PlayerCapsule object for all our player-related work to go in.

Nothing is currently functional, it's just a shell for us to add to.